+++ src/trtutils/__init__.py
@@ -134,7 +134,7 @@ if level is not None and level.upper() not in [
     _log.warning(f"Invalid log level: {level}. Using default log level: WARNING")
 
 __author__ = "Justin Davis"
-__version__ = "0.3.5"
+__version__ = "0.4.0"
 
 import contextlib